+++ src/base/events.h
@@ -274,6 +274,11 @@ typedef enum _event_type_t {
    */
   EVT_DRAG_END,
   /**
+   * @const EVT_SCREEN_SAVER
+   * 在指定的时间内(WITH_SCREEN_SAVER_TIME)，没有用户输入事件，由窗口管理器触发。
+   */
+  EVT_SCREEN_SAVER,
+  /**
    * @const EVT_REQ_START
    * event queue其它请求编号起始值。
    */